move functions to Ui::WindowManager
